Add schema validator
The schema validator will iterate over the various identity fixtures for each gateway
and identity type to validate each sample identity against the corresponding schema.

Any errors are collected during the run and reported back at the end of the run,
with an error exit code. This could be used for CI/CD, PR checks, etc., to pass/fail
changes to identities.