refactor to support arbitrary files
Refactor the file loading logic to assume less about the filesystem
format.  The schema file must be provided via the -schema flag. Any
additional arguments on the command line are assumed to be documents and
are loaded sequentially before being validated against the given schema.
The special file name "-" is also supported; if present, the document
content is loaded for stdin enabling output to be piped into this
command for validation.